requirements.txt
fastapi
uvicorn[standard]
easyocr
opencv-python-headless
pillow
numpy
python-multipart
jinja2
torch==1.13.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
